% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_site}
\alias{build_site}
\title{Build a complete 'pkgdown' website}
\usage{
build_site(pkg = ".", ..., github = TRUE, preview = NA)
}
\arguments{
\item{pkg}{Path to package.}

\item{github}{If TRUE, GitHub specific files are created, e.g. CNAME.}

\item{preview}{Preview site in browser?}

\item{\ldots}{... Additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.}
}
\value{
Nothing.
}
\description{
Build a complete 'pkgdown' website
}
\details{
This function is an enhancement to \code{\link[pkgdown:build_site]{pkgdown::build_site()}} with the
same name.
}
