% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_build.R
\name{pkgdown_build_site}
\alias{pkgdown_build_site}
\title{Build a complete 'pkgdown' website}
\usage{
pkgdown_build_site(pkg = ".", ..., preview = NA)
}
\arguments{
\item{pkg}{Path to package.}

\item{preview}{Preview site in browser?}

\item{\ldots}{... Additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.}
}
\value{
Nothing.
}
\description{
Build a complete 'pkgdown' website
}
