% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_news_md.R
\name{prune_news_md}
\alias{prune_news_md}
\title{Converts an NEWS.md file into a pkgdown formatted NEWS.md file}
\usage{
prune_news_md(pkg = ".", input = "NEWS.md", output = NULL, package = NULL)
}
\arguments{
\item{pkg}{Package path.}

\item{input}{The name of the input \file{NEWS.md} file.}

\item{output}{The name of the output \file{NEWS.md} file.}

\item{package}{The name of the package. If not specified, it is inferred
from the \file{DESCRIPTION} file.}
}
\value{
(invisible) The Markdown output.
}
\description{
Converts an NEWS.md file into a pkgdown formatted NEWS.md file
}
