% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_to_md.R
\name{news_to_md}
\alias{news_to_md}
\title{Converts a traditional NEWS files into a Markdown NEWS.md File}
\usage{
news_to_md(
  pkg = ".",
  input = "NEWS",
  output = "NEWS.md",
  overwrite = FALSE,
  package = NULL,
  style = c("NEWS", "pkgdown"),
  category_case = c("TitleCase", "as-is"),
  escape = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{pkg}{Package path.}

\item{input}{The name of the input \file{NEWS} file.}

\item{output}{The name of the output \file{NEWS.md} file.}

\item{overwrite}{If TRUE, an existing output file is overwritten.}

\item{package}{The name of the package. If not specified, it is inferred
from the \file{DESCRIPTION} file.}

\item{style}{The style of the \file{NEWS.md} file.}

\item{category_case}{The style of the category headers.}

\item{escape}{If TRUE, symbols that have special meaning in Markdown are
escaped.}

\item{verbose}{If TRUE, a summary of the parsed \file{NEWS} file is
outputted.  This can be useful to identify mistypes category titles.}
}
\value{
(invisible) The Markdown output.
}
\description{
Converts a traditional NEWS files into a Markdown NEWS.md File
}
